Add dropdown to variable controls, renamed component
This uesed to be the "Sliders" component, but it does more than
offer slides so a rename was in order!
