+++ src/components/report/VariableControls.css
@@ -47,27 +47,28 @@
 	pointer-events: none;
 }
 
-.instances li {
+/* Instances as button list */
+.instances-list li {
 	display: grid;
 	grid-template-columns: repeat(auto-fill, minmax(10rem, 1fr));
 	grid-gap: 0.25rem;
 }
 
-.instances .button {
+.instances-list .button {
 	background: var(--yellow);
 }
 
-.instances .button.active {
+.instances-list .button.active {
 	background: var(--green);
 }
 
-.instances .button > span {
+.instances-list .button > span {
 	display: flex;
 	align-items: center;
 	text-align: left;
 }
 
-.sample {
+.instances-list .sample {
 	flex: none;
 	font-family: wakamai-fondue, monospace;
 	font-size: 2.5em;
