+++ src/components/report/VariableControls.js
